refactor(core): extract pathkit layer and collapse the pathops adapter
Move everything that calls PathKit into src/core/pathkit: the wasm
loader (loadPathKit), the PathKit typings, the verb-level contour maths
and convertEvenoddToWinding. The flattener's pathops module now drives
PathKit directly and returns SVG command sequences, replacing the
Python-shaped buildPathopsBackend facade and the adapter that translated
its verbs back. Contour ordering and start-point canonicalisation are
unchanged. Drops the test-only PathOps.difference and boundingBox.
src/core/svg/flatten moves to src/core/flatten.
