+++ packages/react-native-nano-icons/src/core/flatten/arcs.ts
