+++ packages/react-native-nano-icons/src/core/flatten/dom.ts
