+++ packages/react-native-nano-icons/src/core/flatten/geometry.ts
