+++ packages/react-native-nano-icons/src/core/flatten/index.ts
@@ -1,0 +1,13 @@
+import type { PathKitModule } from '../pathkit/types';
+import { createPathOps } from './pathops';
+import { PicoSVG } from './svg';
+
+export function flattenSvg(svgContent: string, pathkit: PathKitModule): string {
+  const ops = createPathOps(pathkit);
+  const svg = PicoSVG.fromString(svgContent, ops);
+  svg.topicosvg();
+  return svg.toString();
+}
+
+export { PicoSVG } from './svg';
+export { createPathOps, PathOpsError, type PathOps } from './pathops';
