+++ packages/react-native-nano-icons/src/core/flatten/inherit.ts
