+++ packages/react-native-nano-icons/src/core/flatten/path.ts
