+++ packages/react-native-nano-icons/src/core/flatten/shape.ts
