+++ packages/react-native-nano-icons/src/core/flatten/svg.ts
