+++ packages/react-native-nano-icons/src/core/flatten/transform.ts
